Removed stddef.h

+++ 0x13-more_singly_linked_lists/0-print_listint.c
@@ -22,5 +22,5 @@ size_t print_listint(const listint_t *h)
 		h = h->next;
 	}
 
-	return ((size_t) nodes);
+	return (nodes);
 }
